+++ wetterdienst/provider/imgw/hydrology/api.py
@@ -8,13 +8,14 @@ import re
 from concurrent.futures import ThreadPoolExecutor
 from enum import Enum
 from io import StringIO
+from typing import TYPE_CHECKING
 
 import polars as pl
 import portion as P
 from dateutil.relativedelta import relativedelta
 from fsspec.implementations.zip import ZipFileSystem
 
-from wetterdienst import Kind, Period, Provider, Resolution, Settings
+from wetterdienst import Kind, Parameter, Period, Provider, Resolution, Settings
 from wetterdienst.core.timeseries.request import TimeseriesRequest
 from wetterdienst.core.timeseries.values import TimeseriesValues
 from wetterdienst.metadata.columns import Columns
@@ -27,6 +28,9 @@ from wetterdienst.util.cache import CacheExpiry
 from wetterdienst.util.geo import convert_dms_string_to_dd
 from wetterdienst.util.network import download_file, list_remote_files_fsspec
 from wetterdienst.util.parameter import DatasetTreeCore
+
+if TYPE_CHECKING:
+    from collections.abc import Sequence
 
 log = logging.getLogger(__name__)
 
@@ -340,8 +344,8 @@ class ImgwHydrologyRequest(TimeseriesRequest):
 
     def __init__(
         self,
-        parameter,
-        resolution,
+        parameter: str | ImgwHydrologyParameter | Parameter | Sequence[str | ImgwHydrologyParameter | Parameter],
+        resolution: str | ImgwHydrologyResolution | Resolution,
         start_date: str | dt.datetime | None = None,
         end_date: str | dt.datetime | None = None,
         settings: Settings | None = None,