+++ wetterdienst/provider/wsv/pegel/api.py
@@ -24,6 +24,8 @@ from wetterdienst.util.network import download_file
 from wetterdienst.util.parameter import DatasetTreeCore
 
 if TYPE_CHECKING:
+    from collections.abc import Sequence
+
     from wetterdienst.metadata.parameter import Parameter
     from wetterdienst.settings import Settings
 
@@ -223,7 +225,7 @@ class WsvPegelRequest(TimeseriesRequest):
 
     def __init__(
         self,
-        parameter: list[str | WsvPegelParameter | Parameter],
+        parameter: str | WsvPegelParameter | Parameter | Sequence[str | WsvPegelParameter | Parameter],
         start_date: str | dt.datetime | None = None,
         end_date: str | dt.datetime | None = None,
         settings: Settings | None = None,